+++ src/core/logger/mod.rs
@@ -1,17 +1,3 @@
-pub struct Tracer;
-
-impl Tracer {
-    pub fn new() -> Self {
-        Tracer
-    }
-
-    pub fn trace(&self, message: &str) {
-        println!("[TRACING] {}", message);
-    }
-}
-
-impl Default for Tracer {
-    fn default() -> Self {
-        Self::new()
-    }
-}
+/// Re-export Tracer from the logging module for backward compatibility.
+/// New code should use `crate::logging::Tracer` or `crate::logging::tracer::Tracer`.
+pub use crate::logging::tracer::Tracer;
