+++ src/logging/tracer.rs
@@ -1,0 +1,18 @@
+/// Simple tracer for basic message tracing.
+pub struct Tracer;
+
+impl Tracer {
+    pub fn new() -> Self {
+        Tracer
+    }
+
+    pub fn trace(&self, message: &str) {
+        println!("[TRACING] {}", message);
+    }
+}
+
+impl Default for Tracer {
+    fn default() -> Self {
+        Self::new()
+    }
+}
